Rename internal to placement
