+++ pkg/actors/reminders/reminders.go
@@ -1,0 +1,26 @@
+/*
+Copyright 2023 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package reminders
+
+import (
+	"context"
+)
+
+// RemindersProvider is the interface for the object that provides reminders services.
+type RemindersProvider interface {
+	GetReminder(ctx context.Context, req *GetReminderRequest) (*Reminder, error)
+	CreateReminder(ctx context.Context, req *CreateReminderRequest) error
+	DeleteReminder(ctx context.Context, req *DeleteReminderRequest) error
+	RenameReminder(ctx context.Context, req *RenameReminderRequest) error
+}
